fix update script
